Avoid using the globstar option in bash
When testing an android app with maestro, the bash script generated relies the globstar option. This is only available on bash 4 or higher but mac os only comes with bash 3 (and will probably never move to bash 4 or 5 because they use the gplv3 license). Because of this if you run `eas -p android --local` on a mac the build will fail with this error:

    shopt: globstar: invalid shell option name

To resolve this, I've refactored the maestro test code to use `find` rather than globstar. This meant changing the interface to the `createBuildStepsFromFunctionGroupCall` function slightly (to take a directory rather than a glob expression) but this doesn't seem to cause any issues based on my own testing.